Restores prometheus direct messaging and cleanUp function
