+++ src/firebase/config.js
@@ -9,6 +9,6 @@ const firebaseConfig = {
    appId: process.env.APP_ID,
    measurementId: process.env.MEASUREMENT_ID
 };
-exports.firebaseConfig = firebaseConfig;
+module.exports.firebaseConfig = firebaseConfig;
